requirements.txt
flask
rembg
Pillow
onnxruntime
numpy
